src/third_party/blink/web_tests/external/wpt/css/css-values/minmax-length-percent-serialize-expected.txt

This is a testharness.js-based test.
Found 60 tests; 22 PASS, 38 FAIL, 0 TIMEOUT, 0 NOTRUN.
FAIL e.style['margin-left'] = "min(1px + 1%)" should set the property value assert_equals: serialization should be canonical expected "min(1% + 1px)" but got "min(1px + 1%)"
FAIL e.style['margin-left'] = "min(1cm + 1%)" should set the property value assert_equals: serialization should be canonical expected "min(1% + 1cm)" but got "min(1cm + 1%)"
FAIL e.style['margin-left'] = "min(1mm + 1%)" should set the property value assert_equals: serialization should be canonical expected "min(1% + 1mm)" but got "min(1mm + 1%)"
FAIL e.style['margin-left'] = "min(1Q + 1%)" should set the property value assert_equals: serialization should be canonical expected "min(1% + 1q)" but got "min(1q + 1%)"
FAIL e.style['margin-left'] = "min(1in + 1%)" should set the property value assert_equals: serialization should be canonical expected "min(1% + 1in)" but got "min(1in + 1%)"
FAIL e.style['margin-left'] = "min(1pc + 1%)" should set the property value assert_equals: serialization should be canonical expected "min(1% + 1pc)" but got "min(1pc + 1%)"
FAIL e.style['margin-left'] = "min(1pt + 1%)" should set the property value assert_equals: serialization should be canonical expected "min(1% + 1pt)" but got "min(1pt + 1%)"
FAIL e.style['margin-left'] = "min(1em + 1%)" should set the property value assert_equals: serialization should be canonical expected "min(1% + 1em)" but got "min(1em + 1%)"
FAIL e.style['margin-left'] = "min(1ex + 1%)" should set the property value assert_equals: serialization should be canonical expected "min(1% + 1ex)" but got "min(1ex + 1%)"
FAIL e.style['margin-left'] = "min(1ch + 1%)" should set the property value assert_equals: serialization should be canonical expected "min(1% + 1ch)" but got "min(1ch + 1%)"
FAIL e.style['margin-left'] = "min(1rem + 1%)" should set the property value assert_equals: serialization should be canonical expected "min(1% + 1rem)" but got "min(1rem + 1%)"
FAIL e.style['margin-left'] = "min(1vh + 1%)" should set the property value assert_equals: serialization should be canonical expected "min(1% + 1vh)" but got "min(1vh + 1%)"
FAIL e.style['margin-left'] = "min(1vw + 1%)" should set the property value assert_equals: serialization should be canonical expected "min(1% + 1vw)" but got "min(1vw + 1%)"
FAIL e.style['margin-left'] = "min(1vmin + 1%)" should set the property value assert_equals: serialization should be canonical expected "min(1% + 1vmin)" but got "min(1vmin + 1%)"
FAIL e.style['margin-left'] = "min(1vmax + 1%)" should set the property value assert_equals: serialization should be canonical expected "min(1% + 1vmax)" but got "min(1vmax + 1%)"
FAIL e.style['margin-left'] = "max(1px + 1%)" should set the property value assert_equals: serialization should be canonical expected "max(1% + 1px)" but got "max(1px + 1%)"
FAIL e.style['margin-left'] = "max(1cm + 1%)" should set the property value assert_equals: serialization should be canonical expected "max(1% + 1cm)" but got "max(1cm + 1%)"
FAIL e.style['margin-left'] = "max(1mm + 1%)" should set the property value assert_equals: serialization should be canonical expected "max(1% + 1mm)" but got "max(1mm + 1%)"
FAIL e.style['margin-left'] = "max(1Q + 1%)" should set the property value assert_equals: serialization should be canonical expected "max(1% + 1q)" but got "max(1q + 1%)"
FAIL e.style['margin-left'] = "max(1in + 1%)" should set the property value assert_equals: serialization should be canonical expected "max(1% + 1in)" but got "max(1in + 1%)"
FAIL e.style['margin-left'] = "max(1pc + 1%)" should set the property value assert_equals: serialization should be canonical expected "max(1% + 1pc)" but got "max(1pc + 1%)"
FAIL e.style['margin-left'] = "max(1pt + 1%)" should set the property value assert_equals: serialization should be canonical expected "max(1% + 1pt)" but got "max(1pt + 1%)"
FAIL e.style['margin-left'] = "max(1em + 1%)" should set the property value assert_equals: serialization should be canonical expected "max(1% + 1em)" but got "max(1em + 1%)"
FAIL e.style['margin-left'] = "max(1ex + 1%)" should set the property value assert_equals: serialization should be canonical expected "max(1% + 1ex)" but got "max(1ex + 1%)"
FAIL e.style['margin-left'] = "max(1ch + 1%)" should set the property value assert_equals: serialization should be canonical expected "max(1% + 1ch)" but got "max(1ch + 1%)"
FAIL e.style['margin-left'] = "max(1rem + 1%)" should set the property value assert_equals: serialization should be canonical expected "max(1% + 1rem)" but got "max(1rem + 1%)"
FAIL e.style['margin-left'] = "max(1vh + 1%)" should set the property value assert_equals: serialization should be canonical expected "max(1% + 1vh)" but got "max(1vh + 1%)"
FAIL e.style['margin-left'] = "max(1vw + 1%)" should set the property value assert_equals: serialization should be canonical expected "max(1% + 1vw)" but got "max(1vw + 1%)"
FAIL e.style['margin-left'] = "max(1vmin + 1%)" should set the property value assert_equals: serialization should be canonical expected "max(1% + 1vmin)" but got "max(1vmin + 1%)"
FAIL e.style['margin-left'] = "max(1vmax + 1%)" should set the property value assert_equals: serialization should be canonical expected "max(1% + 1vmax)" but got "max(1vmax + 1%)"
PASS e.style['margin-left'] = "min(20px, 10%)" should set the property value
PASS e.style['margin-left'] = "min(1em, 10%)" should set the property value
PASS e.style['margin-left'] = "max(20px, 10%)" should set the property value
PASS e.style['margin-left'] = "max(1em, 10%)" should set the property value
PASS e.style['margin-left'] = "min(10%, 20px)" should set the property value
PASS e.style['margin-left'] = "min(10%, 1em)" should set the property value
PASS e.style['margin-left'] = "max(10%, 20px)" should set the property value
PASS e.style['margin-left'] = "max(10%, 1em)" should set the property value
PASS e.style['margin-left'] = "min(10% + 30px, 5% + 60px)" should set the property value
PASS e.style['margin-left'] = "max(10% + 2em, 5% + 1em)" should set the property value
PASS e.style['margin-left'] = "calc(min(10%) + max(1em) + min(20px))" should set the property value
PASS e.style['margin-left'] = "calc(max(20px) + min(1em) + max(10%))" should set the property value
PASS e.style['margin-left'] = "calc(max(10%) + min(1em) + max(20px))" should set the property value
PASS e.style['margin-left'] = "calc(min(20px) + max(1em) + min(10%))" should set the property value
PASS e.style['margin-left'] = "calc(20px + min(10%))" should set the property value
PASS e.style['margin-left'] = "calc(10% + min(20px))" should set the property value
PASS e.style['margin-left'] = "calc(1em + min(10%))" should set the property value
PASS e.style['margin-left'] = "calc(10% + min(1em))" should set the property value
FAIL e.style['margin-left'] = "calc(min(10%) + 20px)" should set the property value assert_equals: serialization should be canonical expected "calc(20px + min(10%))" but got "calc(min(10%) + 20px)"
FAIL e.style['margin-left'] = "calc(min(20px) + 10%)" should set the property value assert_equals: serialization should be canonical expected "calc(10% + min(20px))" but got "calc(min(20px) + 10%)"
FAIL e.style['margin-left'] = "calc(min(10%) + 1em)" should set the property value assert_equals: serialization should be canonical expected "calc(1em + min(10%))" but got "calc(min(10%) + 1em)"
FAIL e.style['margin-left'] = "calc(min(1em) + 10%)" should set the property value assert_equals: serialization should be canonical expected "calc(10% + min(1em))" but got "calc(min(1em) + 10%)"
PASS e.style['margin-left'] = "calc(20px + max(10%))" should set the property value
PASS e.style['margin-left'] = "calc(10% + max(20px))" should set the property value
PASS e.style['margin-left'] = "calc(1em + max(10%))" should set the property value
PASS e.style['margin-left'] = "calc(10% + max(1em))" should set the property value
FAIL e.style['margin-left'] = "calc(max(10%) + 20px)" should set the property value assert_equals: serialization should be canonical expected "calc(20px + max(10%))" but got "calc(max(10%) + 20px)"
FAIL e.style['margin-left'] = "calc(max(20px) + 10%)" should set the property value assert_equals: serialization should be canonical expected "calc(10% + max(20px))" but got "calc(max(20px) + 10%)"
FAIL e.style['margin-left'] = "calc(max(10%) + 1em)" should set the property value assert_equals: serialization should be canonical expected "calc(1em + max(10%))" but got "calc(max(10%) + 1em)"
FAIL e.style['margin-left'] = "calc(max(1em) + 10%)" should set the property value assert_equals: serialization should be canonical expected "calc(10% + max(1em))" but got "calc(max(1em) + 10%)"
Harness: the test ran to completion.
